+++ uniffi_bindgen/src/bindings/python/templates/ObjectTemplate.py
@@ -5,9 +5,12 @@
 
 {% include "Protocol.py" %}
 
+{% if ci.is_name_used_as_error(name) %}
+class {{ impl_name }}(Exception):
+{%- else %}
 class {{ impl_name }}:
+{%- endif %}
     {%- call py::docstring(obj, 4) %}
-
     _pointer: ctypes.c_void_p
 
 {%- match obj.primary_constructor() %}
@@ -53,8 +56,7 @@ class {{ impl_name }}:
 
 {%- for meth in obj.methods() -%}
     {%- call py::method_decl(meth.name()|fn_name, meth) %}
-{% endfor %}
-
+{%- endfor %}
 {%- for tm in obj.uniffi_traits() -%}
 {%-     match tm %}
 {%-         when UniffiTrait::Debug { fmt } %}
@@ -75,14 +77,38 @@ class {{ impl_name }}:
         return {{ ne.return_type().unwrap()|lift_fn }}({% call py::to_ffi_call_with_prefix("self._uniffi_clone_pointer()", ne) %})
 {%-         when UniffiTrait::Hash { hash } %}
             {%- call py::method_decl("__hash__", hash) %}
-{%      endmatch %}
-{% endfor %}
+{%-      endmatch %}
+{%- endfor %}
 
 {%- if obj.has_callback_interface() %}
 {%- let ffi_init_callback = obj.ffi_init_callback() %}
 {%- let vtable = obj.vtable().expect("trait interface should have a vtable") %}
 {%- let vtable_methods = obj.vtable_methods() %}
 {% include "CallbackInterfaceImpl.py" %}
+{%- endif %}
+
+{# Objects as error #}
+{%- if ci.is_name_used_as_error(name) %}
+{# Due to some mismatches in the ffi converter mechanisms, errors are forced to be a RustBuffer #}
+class {{ ffi_converter_name }}__as_error(_UniffiConverterRustBuffer):
+    @classmethod
+    def read(cls, buf):
+        raise NotImplementedError()
+
+    @classmethod
+    def write(cls, value, buf):
+        raise NotImplementedError()
+
+    @staticmethod
+    def lift(value):
+        # Errors are always a rust buffer holding a pointer - which is a "read"
+        with value.consume_with_stream() as stream:
+            return {{ ffi_converter_name }}.read(stream)
+
+    @staticmethod
+    def lower(value):
+        raise NotImplementedError()
+
 {%- endif %}
 
 class {{ ffi_converter_name }}: