+++ uniffi_bindgen/src/bindings/python/templates/macros.py
@@ -5,27 +5,29 @@
 #}
 
 {%- macro to_ffi_call(func) -%}
-    {%- match func.throws_type() -%}
-    {%- when Some with (e) -%}
-_rust_call_with_error({{ e|ffi_converter_name }},
-    {%- else -%}
-_rust_call(
-    {%- endmatch -%}
-    _UniffiLib.{{ func.ffi_func().name() }},
-    {%- call arg_list_lowered(func) -%}
-)
+{%- call _to_ffi_call_with_prefix_arg("", func) %}
 {%- endmacro -%}
 
 {%- macro to_ffi_call_with_prefix(prefix, func) -%}
-    {%- match func.throws_type() -%}
-    {%- when Some with (e) -%}
-_rust_call_with_error(
-    {{ e|ffi_converter_name }},
-    {%- else -%}
+{%- call _to_ffi_call_with_prefix_arg(format!("{},", prefix), func) %}
+{%- endmacro -%}
+
+{%- macro _to_ffi_call_with_prefix_arg(prefix, func) -%}
+{%- match func.throws_type() -%}
+{%-     when Some with (e) -%}
+{%-         match e -%}
+{%-             when Type::Enum { name, module_path } -%}
+_rust_call_with_error({{ e|ffi_converter_name }},
+{%-             when Type::Object { name, module_path, imp } -%}
+_rust_call_with_error({{ e|ffi_converter_name }}__as_error,
+{%-             else %}
+# unsupported error type!
+{%-         endmatch %}
+{%- else -%}
 _rust_call(
-    {%- endmatch -%}
+{%- endmatch -%}
     _UniffiLib.{{ func.ffi_func().name() }},
-    {{- prefix }},
+    {{- prefix }}
     {%- call arg_list_lowered(func) -%}
 )
 {%- endmacro -%}
@@ -138,13 +140,7 @@ _rust_call(
             {%- when None %}
             lambda val: None,
             {% endmatch %}
-            # Error FFI converter
-            {%- match meth.throws_type() %}
-            {%- when Some(e) %}
-            {{ e|ffi_converter_name }},
-            {%- when None %}
-            None,
-            {%- endmatch %}
+            {% call error_ffi_converter(meth) %}
         )
 
 {%- else -%}
@@ -168,4 +164,21 @@ _rust_call(
 {%      endmatch %}
 {%  endif %}
 
+{% endmacro %}
+
+{%- macro error_ffi_converter(func) %}
+    # Error FFI converter
+{%  match func.throws_type() %}
+{%-     when Some(e) %}
+{%-         match e -%}
+{%-             when Type::Enum { name, module_path } -%}
+    {{ e|ffi_converter_name }},
+{%-             when Type::Object { name, module_path, imp } -%}
+    {{ e|ffi_converter_name }}__as_error,
+{%-             else %}
+    # unsupported error type!
+{%-         endmatch %}
+{%-     when None %}
+    None,
+{%-  endmatch %}
 {% endmacro %}